Update button position
