+++ public/app/plugins/datasource/tempo/SearchTraceQLEditor/TraceQLSearch.tsx
@@ -1,10 +1,9 @@
 import { css } from '@emotion/css';
 import React, { useCallback, useEffect, useState } from 'react';
 
-import { GrafanaTheme2 } from '@grafana/data';
-import { EditorRow } from '@grafana/experimental';
-import { config, FetchError, getTemplateSrv } from '@grafana/runtime';
-import { Alert, HorizontalGroup, useStyles2 } from '@grafana/ui';
+import { CoreApp, GrafanaTheme2 } from '@grafana/data';
+import { config, FetchError, getTemplateSrv, reportInteraction } from '@grafana/runtime';
+import { Alert, Button, HorizontalGroup, useStyles2 } from '@grafana/ui';
 
 import { createErrorNotification } from '../../../../core/copy/appNotification';
 import { notifyApp } from '../../../../core/reducers/appNotification';
@@ -28,11 +27,13 @@ interface Props {
   query: TempoQuery;
   onChange: (value: TempoQuery) => void;
   onBlur?: () => void;
+  onClearResults: () => void;
+  app?: CoreApp;
 }
 
 const hardCodedFilterIds = ['min-duration', 'max-duration', 'status'];
 
-const TraceQLSearch = ({ datasource, query, onChange }: Props) => {
+const TraceQLSearch = ({ datasource, query, onChange, onClearResults, app }: Props) => {
   const styles = useStyles2(getStyles);
   const [error, setError] = useState<Error | FetchError | null>(null);
 
@@ -196,9 +197,32 @@ const TraceQLSearch = ({ datasource, query, onChange }: Props) => {
             <GroupByField datasource={datasource} onChange={onChange} query={query} isTagsLoading={isTagsLoading} />
           )}
         </div>
-        <EditorRow>
+        <div className={styles.rawQueryContainer}>
           <RawQuery query={templateSrv.replace(traceQlQuery)} lang={{ grammar: traceqlGrammar, name: 'traceql' }} />
-        </EditorRow>
+          {query.queryType === 'traceqlSearch' && (
+            <Button
+              variant="secondary"
+              size="sm"
+              onClick={() => {
+                reportInteraction('grafana_traces_copy_to_traceql_clicked', {
+                  app: app ?? '',
+                  grafana_version: config.buildInfo.version,
+                  location: 'search_tab',
+                });
+
+                onClearResults();
+                const traceQlQuery = generateQueryFromFilters(query.filters || []);
+                onChange({
+                  ...query,
+                  query: traceQlQuery,
+                  queryType: 'traceql',
+                });
+              }}
+            >
+              Copy to TraceQL
+            </Button>
+          )}
+        </div>
         <TempoQueryBuilderOptions onChange={onChange} query={query} />
       </div>
       {error ? (
@@ -224,4 +248,11 @@ const getStyles = (theme: GrafanaTheme2) => ({
     flex-wrap: wrap;
     flex-direction: column;
   `,
+  rawQueryContainer: css({
+    alignItems: 'center',
+    backgroundColor: theme.colors.background.secondary,
+    display: 'flex',
+    justifyContent: 'space-between',
+    padding: theme.spacing(1),
+  }),
 });
